refactor: address PR review feedback — extract subcomponents, scope zoom, rename !lag to -lag
- Replace `!lag` with `-lag` in all 4 occurrences (reviewer feedback)
- Move chartjs-plugin-zoom from global registration to per-instance plugin
  so zoom only affects the lag report chart, not OverallStatistics graphs
- Extract AdminLagReportDetail (1182 lines) into 6 focused subcomponents:
  LagReportHeader, LagReportTopology, LagReportContinuousMonitoring,
  LagReportInspector, LagReportEventData, LagReportRawData
- Parent component reduced to 123-line coordination shell
