+++ src/components/overall-statistics/BarChart.vue
@@ -3,13 +3,13 @@
     ref="chartRef"
     :data="chartData"
     :options="chartOptions"
+    :plugins="chartPlugins"
   />
 </template>
 
 <script lang="ts">
 import { BarController, BarElement, CategoryScale, Chart as ChartJS, ChartOptions, Filler, LinearScale, Tooltip, Legend, ChartData } from "chart.js";
 import chartJSPluginAnnotation from "chartjs-plugin-annotation";
-import zoomPlugin from "chartjs-plugin-zoom";
 import { PropType } from "vue";
 import { Bar as BarChartGeneric } from "vue-chartjs";
 
@@ -20,7 +20,6 @@ ChartJS.register(LinearScale);
 ChartJS.register(Filler);
 ChartJS.register(Tooltip);
 ChartJS.register(chartJSPluginAnnotation);
-ChartJS.register(zoomPlugin);
 ChartJS.register(Legend);
 
 const defaultOptions = (): ChartOptions => {
@@ -66,6 +65,10 @@ export default {
     chartOptions: {
       type: Object,
       default: defaultOptions,
+    },
+    chartPlugins: {
+      type: Array,
+      default: (): unknown[] => [],
     },
     // datasetIdKey: {
     //   type: String,
